Modifier WIP
I am commiting the WIP Modifier GlobalItem for handling adding modifiers and a GlobalNPC script for the corresponding on-hit effects. There is also a new GlobalProjectile script that is used to track the projectile owner for applying the proper on-hit effects.

I moved some system scripts to the correct folder since they were Global scripts but was in the Content folder instead of Common.