Initial Commit
The main purpose of this commit is the initial version of the skill system; Has a variety of other additions as well. Added a GlobalItems script for performing generalized changes to items, like guns not using ammo and tooltip cleanup. Also, added a GlobalTiles script for preventing tiles from dropping items and explosions from destroying any tiles. Some small optimizations for existing systems, reworked on-hit effects based on feedback ,and added localization.